requirements.txt
pandas
mysql-connector-python
transformers
torch
streamlit
seaborn
matplotlib
google-api-python-client
